// server/src/types/employee.ts
export interface Employee {
    Employee_Name: string;
    Employee_EmailID: string;
  }
  
  export interface PreviousAssignment {
    Employee_Name: string;
    Employee_EmailID: string;
    Secret_Child_Name: string;
    Secret_Child_EmailID: string;
  }
  
  export interface EmployeeAssignment {
    Employee_Name: string;
    Employee_EmailID: string;
    Secret_Child_Name: string;
    Secret_Child_EmailID: string;
  }